refactor settings screen
