+++ lib/components/settings_title.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class SettingsTitle extends StatelessWidget {
+  const SettingsTitle({Key? key, required this.text}) : super(key: key);
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(
+        right: 20.0,
+      ),
+      child: Text(
+        text,
+        textAlign: TextAlign.start,
+        style: Theme.of(context).textTheme.subtitle1,
+        textScaleFactor: 1,
+      ),
+    );
+  }
+}
