+++ lib/components/vertical_space.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class VerticalSpace extends StatelessWidget {
+  const VerticalSpace({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox(height: 10.0);
+  }
+}
