+++ lib/exports/components.dart
@@ -1,18 +1,35 @@
-export 'package:noor/components/card_template.dart';
 export 'package:noor/components/allah_names_title.dart';
+export 'package:noor/components/animated_size_text.dart';
 export 'package:noor/components/athkar_card.dart';
 export 'package:noor/components/athkar_title.dart';
 export 'package:noor/components/bottom_nav.dart';
-export 'package:noor/components/home-card.dart';
+export 'package:noor/components/card_template.dart';
+export 'package:noor/components/card_text.dart';
 export 'package:noor/components/close_button.dart';
-export 'package:noor/components/delete_dialog.dart';
+export 'package:noor/components/copy_action.dart';
 export 'package:noor/components/custom_scroll_bhaviour.dart';
+export 'package:noor/components/delete_dialog.dart';
+export 'package:noor/components/fav_action.dart';
+export 'package:noor/components/font_type_button.dart';
 export 'package:noor/components/glowing_stars.dart';
+export 'package:noor/components/home-card.dart';
 export 'package:noor/components/image_button.dart';
 export 'package:noor/components/list_item.dart';
 export 'package:noor/components/logo.dart';
+export 'package:noor/components/noor_card.dart';
 export 'package:noor/components/noor_icons.dart';
 export 'package:noor/components/noor_settings_icons.dart';
-export 'package:noor/components/card_text.dart';
-export 'package:noor/components/copy_action.dart';
-export 'package:noor/components/fav_action.dart';
+export 'package:noor/components/radio_btn.dart';
+export 'package:noor/components/segmented_control_option.dart';
+export 'package:noor/components/settings/about.dart';
+export 'package:noor/components/settings/appearance.dart';
+export 'package:noor/components/settings/counter.dart';
+export 'package:noor/components/settings/line.dart';
+export 'package:noor/components/settings/notifications.dart';
+export 'package:noor/components/settings/sources.dart';
+export 'package:noor/components/settings/vibration.dart';
+export 'package:noor/components/settings_subtitle.dart';
+export 'package:noor/components/settings_title.dart';
+export 'package:noor/components/subtitle_with_icon.dart';
+export 'package:noor/components/switcher_option.dart';
+export 'package:noor/components/vertical_space.dart';
